refactor : add formatDateOfBirth to actor details
