+++ gradle/libs.versions.toml
@@ -7,7 +7,6 @@ coreKtx = "1.6.0"
 kotlin = "1.9.10"
 compose = "1.5.3"
 composeBom = "2023.10.00"
-kotlinxDatetime = "0.5.0"
 ksp = "1.9.10-1.0.13"
 google-services = "4.4.0"
 firebase-crashlytics = "2.9.9"
@@ -20,7 +19,7 @@ retrofit = "2.11.0"
 room = "2.7.2"
 ktor = "2.3.2"
 kotlinx-serialization = "1.6.0"
-kotlinx-datetime = "0.4.1"
+kotlinx-datetime = "0.5.0"
 material3Android = "1.3.2"
 foundationAndroid = "1.5.3"
 foundationAndroidVersion = "1.8.3"
@@ -55,10 +54,9 @@ koin-androidx-compose = { group = "io.insert-koin", name = "koin-androidx-compos
 koin-annotations = { group = "io.insert-koin", name = "koin-annotations", version = "1.3.0" }
 koin-android = { group = "io.insert-koin", name = "koin-android", version = "3.5.0" }
 androidx-appcompat = { group = "androidx.appcompat", name = "appcompat", version = "1.6.1" }
-kotlinx-datetime-v050 = { module = "org.jetbrains.kotlinx:kotlinx-datetime", version.ref = "kotlinxDatetime" }
+kotlinx-datetime = { group = "org.jetbrains.kotlinx", name = "kotlinx-datetime", version.ref = "kotlinx-datetime" }
 kotlinx-serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "kotlinx-serialization" }
 kotlinx-serialization-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-core", version.ref = "kotlinx-serialization" }
-kotlinx-datetime = { group = "org.jetbrains.kotlinx", name = "kotlinx-datetime", version.ref = "kotlinx-datetime" }
 androidx-room-testing = { group = "androidx.room", name = "room-testing", version.ref = "room" }
 androidx-room-paging = { group = "androidx.room", name = "room-paging", version.ref = "room" }
 androidx-work-runtime-ktx = { group = "androidx.work", name = "work-runtime-ktx", version = "2.8.1" }
