+++ presentation/build.gradle.kts
@@ -79,7 +79,7 @@ dependencies {
     //pagination
     implementation(libs.androidx.paging.runtime)
     implementation(libs.androidx.paging.compose)
-    implementation(libs.kotlinx.datetime.v050)
+    implementation(libs.kotlinx.datetime)
     // Serialization
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.kotlinx.serialization.core)
