+++ presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/SeriesDetailsViewModel.kt
@@ -13,10 +13,9 @@ import com.madrid.domain.usecase.series.GetSeriesTopCastUseCase
 import com.madrid.domain.usecase.series.GetSimilarSeriesUseCase
 import com.madrid.presentation.navigation.Destinations
 import com.madrid.presentation.viewModel.base.BaseViewModel
-import com.madrid.presentation.viewModel.uiStateMapper.barser.formatDateKotlinx
+import com.madrid.presentation.viewModel.shared.barser.formatDateKotlinx
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
-import okhttp3.internal.concurrent.formatDuration
 
 class SeriesDetailsViewModel(
     private val savedStateHandle: SavedStateHandle,
